fix: OAuth fallback, float hours parsing, and pre-merge workflow
- Fall back to OAuth flow when token refresh fails instead of crashing
- Fix timesheet parser: handle "Total Logged: N" without h suffix,
  support decimal hours, remove fragile end-of-line fallback pattern
- Change hourly_rate and total_hours to float throughout
- Pre-merge PDFs when all docs arrive (before approval) for review
- Store merged_pdf_path in WorkflowData for send step

+++ src/config.py
@@ -37,7 +37,7 @@ class Settings(BaseSettings):
 
     # Invoice settings
     company_name: str = "YourCompany inc."
-    hourly_rate: int = 10
+    hourly_rate: float = 10
     currency: str = "EUR"
 
     # LLM
